v1: define names for unknown/default iota values
Previously we were using "_", and this wasn't a problem because
protoc-gen-go prefaces enums with the enum name to prevent clashes.
However, this is not the case with other languages (specifically Java).
Give actual names for our unknown/default enum values to prevent protoc
errors during compilation.

While at it, get rid of the ISO 4217 currency enum, as the codes change
frequently and the enum is likely to go out of date.