modify transaction creation to allow offline interaction
Creating a transaction from an offline location, especially for
remittance purposes, was not being handled in the transaction service
definition.

Allows for depositing money from a physical location (atm/branch) and
sending money for physical receipt (remittance transaction).  Note
that the transaction should have a bank account as source or
destination to be valid.