v1: revert some fields to fix the ordering
`ContactNumber` was added in the middle by a previous commit
and introduced some breaking changes for old clients. Reverting
to original position to avoid breaking changes.

Bring back deleted pagination fields for GetAccountsRequest,
GetAccountResponse, GetTransactionsRequest and GetTransactionResponse.
For GetTransactionRequest reverting the data type to string and
adding PageLimit with type int.

This is done to fix breaking changes for old clients.

References:
https://github.com/openbank/openbank/commit/fc34b70677ac7d6563b8ed5d08870e960acfc526
https://github.com/openbank/openbank/commit/399bbe19de2f4b28ab604b01413249a046648f80